backend/auth-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
boto3==1.34.84
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
